+++ Raytracer.jl
@@ -570,7 +570,7 @@ main(x::Union{String, Float64, Int64}...) = main([string(var) for var in [x...]]
 function main(args)
 	parsed_arguments = ArgParse_command_line(args) # the result is a Dict{String,Any}
 	(isnothing(parsed_arguments)) && (return nothing)
-	#print_ArgParseSettings(parsed_arguments)
+	print_ArgParseSettings(parsed_arguments)
 
 	parsed_command = parsed_arguments["%COMMAND%"]
 	parsed_settings = parsed_arguments[parsed_command]